Tests/LibWeb/Text/expected/scroll-to-fragment.txt

Take test     The page has been scrolled to y: 562.59375
